A4A WooPayments: pass classic dashboard data as props like Migrations
The classic WooPayments dashboard reached the shared ConsolidatedViews and
CommissionsTable through a WooPaymentsContext plus per-component adapters,
indirection the Migrations page doesn't have. Follow the Migrations approach:
the page already has the data from useWooPaymentsDashboardData, so pass it as
direct props to the dashboard-content component, which renders the shared
components itself.

Removes the context and both adapter files.
