--- client/a8c-for-agencies/sections/woopayments/context.tsx
@@ -1,28 +1,0 @@
-import { createContext, useContext } from '@wordpress/element';
-import type { WooPaymentsData } from '@automattic/api-core';
-import type { SitesWithWooPaymentsState } from 'calypso/dashboard/agency/earn/woopayments/types';
-
-interface WooPaymentsContextType {
-	woopaymentsData?: WooPaymentsData;
-	isLoadingWooPaymentsData: boolean;
-	sitesWithPluginsStates: SitesWithWooPaymentsState[];
-}
-
-const WooPaymentsContext = createContext< WooPaymentsContextType | undefined >( undefined );
-
-export const useWooPaymentsContext = () => {
-	const context = useContext( WooPaymentsContext );
-	if ( context === undefined ) {
-		throw new Error( 'useWooPaymentsContext must be used within a WooPaymentsProvider' );
-	}
-	return context;
-};
-
-interface WooPaymentsProviderProps {
-	children: React.ReactNode;
-	value: WooPaymentsContextType;
-}
-
-export const WooPaymentsProvider = ( { children, value }: WooPaymentsProviderProps ) => {
-	return <WooPaymentsContext.Provider value={ value }>{ children }</WooPaymentsContext.Provider>;
-};
