+++ client/a8c-for-agencies/sections/woopayments/dashboard-content/index.tsx
@@ -1,13 +1,54 @@
-import WooPaymentsConsolidatedViews from './consolidated-views';
-import SitesWithWooPayments from './sites-with-woopayments';
+import { useCallback } from 'react';
+import CommissionsTable from 'calypso/dashboard/agency/earn/woopayments/commissions-table';
+import ConsolidatedViews from 'calypso/dashboard/agency/earn/woopayments/consolidated-views';
+import { useDownloadCommissionsReport } from 'calypso/dashboard/agency/earn/woopayments/use-download-commissions-report';
+import { useDispatch } from 'calypso/state';
+import { recordTracksEvent } from 'calypso/state/analytics/actions';
+import type { WooPaymentsData } from '@automattic/api-core';
+import type { SitesWithWooPaymentsState } from 'calypso/dashboard/agency/earn/woopayments/types';
 
 import './style.scss';
 
-const WooPaymentsDashboardContent = () => {
+interface WooPaymentsDashboardContentProps {
+	agencyId: number;
+	woopaymentsData?: WooPaymentsData;
+	isLoadingWooPaymentsData: boolean;
+	sitesWithPluginsStates: SitesWithWooPaymentsState[];
+}
+
+const WooPaymentsDashboardContent = ( {
+	agencyId,
+	woopaymentsData,
+	isLoadingWooPaymentsData,
+	sitesWithPluginsStates,
+}: WooPaymentsDashboardContentProps ) => {
+	const dispatch = useDispatch();
+	const { downloadCommissionsReport } = useDownloadCommissionsReport( agencyId );
+
+	const recordTracks = useCallback(
+		( eventName: string, properties?: Record< string, unknown > ) => {
+			dispatch( recordTracksEvent( eventName, properties ) );
+		},
+		[ dispatch ]
+	);
+
 	return (
 		<>
-			<WooPaymentsConsolidatedViews />
-			<SitesWithWooPayments />
+			<div className="consolidated-view">
+				<ConsolidatedViews
+					woopaymentsData={ woopaymentsData }
+					isLoading={ isLoadingWooPaymentsData }
+				/>
+			</div>
+			<div className="redesigned-a8c-table full-width">
+				<CommissionsTable
+					sites={ sitesWithPluginsStates }
+					woopaymentsData={ woopaymentsData }
+					isLoadingWooPaymentsData={ isLoadingWooPaymentsData }
+					recordTracksEvent={ recordTracks }
+					onDownloadReport={ downloadCommissionsReport }
+				/>
+			</div>
 		</>
 	);
 };
